Allocated values outside of loop (fixes GH issue #8).

+++ default_wait_strategy.go
@@ -4,6 +4,6 @@ import "time"
 
 type DefaultWaitStrategy struct{}
 
-func NewWaitStrategy() DefaultWaitStrategy      { return DefaultWaitStrategy{} }
-func (this DefaultWaitStrategy) Gate(count int) { time.Sleep(time.Microsecond) }
-func (this DefaultWaitStrategy) Idle(count int) { time.Sleep(time.Millisecond) }
+func NewWaitStrategy() DefaultWaitStrategy        { return DefaultWaitStrategy{} }
+func (this DefaultWaitStrategy) Gate(count int64) { time.Sleep(time.Microsecond) }
+func (this DefaultWaitStrategy) Idle(count int64) { time.Sleep(time.Millisecond) }
